bower: Change `~` to `^` for `bower >= v1.7.5`
From Polymer Docs

https://www.polymer-project.org/1.0/docs/start/getting-the-code.html

```
Note: Bower versions prior to 1.7.5 defaulted to a narrower version range,
using the tilde operator. For example, #~1.3.1, which matches any version
equal to or greater than 1.3.1, but less than 1.4.0. If you’re using an
older version of Bower, we recommend updating to the latest version and
checking the version ranges in your bower.json files.
```